RavenDB-17223 TimeSeriesTestBase
