+++ test/SlowTests/Client/TimeSeries/Issues/RavenDB-16060.cs
@@ -6,7 +6,6 @@ using FastTests;
 using Raven.Client.Documents;
 using Raven.Client.Documents.Operations.TimeSeries;
 using Raven.Client.Documents.Session;
-using SlowTests.Client.TimeSeries.Query;
 using SlowTests.Core.Utils.Entities;
 using Sparrow;
 using Xunit;
@@ -524,7 +523,7 @@ namespace SlowTests.Client.TimeSeries.Issues
                 await database.TimeSeriesPolicyRunner.RunRollups();
                 await database.TimeSeriesPolicyRunner.DoRetention();
 
-                await QueryFromMultipleTimeSeries.VerifyPolicyExecution(store, config.Collections["Users"], 12, rawName: "StockPrices");
+                await TimeSeries.VerifyPolicyExecution(store, config.Collections["Users"], 12, rawName: "StockPrices");
 
                 using (var session = store.OpenSession())
                 {
@@ -600,7 +599,7 @@ namespace SlowTests.Client.TimeSeries.Issues
                 await database.TimeSeriesPolicyRunner.RunRollups();
                 await database.TimeSeriesPolicyRunner.DoRetention();
 
-                await QueryFromMultipleTimeSeries.VerifyPolicyExecution(store, config.Collections["Users"], 12, rawName: "StockPrices");
+                await TimeSeries.VerifyPolicyExecution(store, config.Collections["Users"], 12, rawName: "StockPrices");
 
                 using (var session = store.OpenSession())
                 {
